+++ core/src/main/scala/slinky/core/facade/Fragment.scala
@@ -3,5 +3,5 @@ package slinky.core.facade
 import slinky.core.ExternalComponentNoProps
 
 object Fragment extends ExternalComponentNoProps {
-  override val component = React.Fragment
+  override val component = ReactRaw.Fragment
 }