+++ tests/src/test/scala/slinky/core/ReactRefTest.scala
@@ -12,7 +12,7 @@ import scala.concurrent.Promise
 
 class ReactRefTest extends AsyncFunSuite {
   test("Can pass in a ref object to an HTML tag and use it") {
-    val elemRef = React.createRef[Element]()
+    val elemRef = React.createRef[Element]
     ReactDOM.render(
       div(ref := elemRef)("hello!"),
       dom.document.createElement("div")
@@ -23,7 +23,7 @@ class ReactRefTest extends AsyncFunSuite {
 
   test("Can pass in a ref object to a Slinky component and use it") {
     val promise: Promise[Assertion] = Promise()
-    val ref = React.createRef[TestForceUpdateComponent.Def]()
+    val ref = React.createRef[TestForceUpdateComponent.Def]
 
     ReactDOM.render(
       TestForceUpdateComponent(() => promise.success(assert(true))).withRef(ref),
@@ -33,5 +33,19 @@ class ReactRefTest extends AsyncFunSuite {
     ref.current.forceUpdate()
 
     promise.future
+  }
+
+  test("Can use forwardRef to pass down a ref to a lower element") {
+    val forwarded = React.forwardRef[String]((props, rf) => {
+      div(ref := rf)(props)
+    })
+
+    val divRef = React.createRef[Any]
+    ReactDOM.render(
+      forwarded("hello").withRef(divRef),
+      dom.document.createElement("div")
+    )
+
+    assert(divRef.current.asInstanceOf[HTMLElement].innerHTML == "hello")
   }
 }